v1.1

+++ backend/database/migrations/2026_04_12_000004_add_quota_fields_to_centres_table.php
@@ -9,9 +9,9 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('centres', function (Blueprint $table) {
-            $table->unsignedTinyInteger('seat_count')->nullable()->after('is_ready');
-            $table->unsignedTinyInteger('short_slots_per_day')->nullable()->after('seat_count');
-            $table->unsignedTinyInteger('long_slots_per_day')->nullable()->after('short_slots_per_day');
+            $table->unsignedSmallInteger('seat_count')->nullable()->after('is_ready');
+            $table->unsignedSmallInteger('short_slots_per_day')->nullable()->after('seat_count');
+            $table->unsignedSmallInteger('long_slots_per_day')->nullable()->after('short_slots_per_day');
         });
     }
